only reed food if table has no entries

+++ CarbTracker.Services/FoodService.cs
@@ -17,8 +17,8 @@ namespace CarbTracker.Services
             using (var context = new ApplicationDbContext())
             {
 
-                //if (context.Foods.Count() < 2)
-                //{
+                if (context.Foods.Count() < 2)
+                {
                     // Warning - reseed regardless of current content !!!
                     context.Database.ExecuteSqlCommand("DELETE FROM dbo.Food"); // DELETE ALL ROWS! 
                     context.Database.ExecuteSqlCommand("DBCC CHECKIDENT('dbo.Food', RESEED, 0)"); // Reseed index
@@ -43,7 +43,7 @@ namespace CarbTracker.Services
                         context.SaveChanges();
                     }
 
-                //}
+                }
                 var query = context.Foods
                                 .Select(e => new FoodListItem
                                 {
